requirements.txt
apache-airflow==2.7.3
sqlalchemy==1.4.51
pandas==2.0.3
matplotlib==3.7.4
seaborn==0.13.0
requests==2.31.0
